# backend/docs/openapi-3.yaml
components:
  schemas:
...
    api.CreatePersonaRequest:
      type: object
...
    api.LoginSuccessResponse:
      properties:
        message:
          type: string
        session:
          $ref: '#/components/schemas/api.SessionData'
        user:
          $ref: '#/components/schemas/api.UserPublicResponse'
      type: object
...
    api.PasswordChangeResponse:
      properties:
        message:
          type: string
        success:
          type: boolean
      type: object
...
    api.SessionRefreshResponse:
      properties:
        expiresAt:
          type: string
        message:
          type: string
      type: object
    api.StandardErrorResponse:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
...
    models.ChangePasswordRequest:
      properties:
        currentPassword:
          type: string
        newPassword:
          minLength: 12
          type: string
      required:
      - currentPassword
      - newPassword
      type: object
...
    models.LoginRequest:
      properties:
        captchaToken:
          type: string
        email:
          type: string
        password:
          minLength: 6
          type: string
        rememberMe:
          type: boolean
      required:
      - email
      - password
      type: object
...
  securitySchemes:
    basic:
      scheme: basic
      type: http
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for domain generation, validation, and management
    operations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Domain Flow API
  version: "1.0"
openapi: 3.1.0
paths:
  /auth/change-password:
    post:
      description: Change password for the currently authenticated user
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.ChangePasswordRequest'
        description: Password change request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PasswordChangeResponse'
          description: Password changed successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid request format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Authentication required
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Not implemented
      summary: Change user password
      tags:
      - authentication
  /auth/login:
    post:
      description: Authenticate user credentials and create session
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.LoginRequest'
        description: Login credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.LoginSuccessResponse'
          description: Login successful with user and session info
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid request format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid credentials
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Account inactive
        "423":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Account locked
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Internal server error
      summary: User login
      tags:
      - authentication
...
  /auth/me:
    get:
      description: Get information about the currently authenticated user
      operationId: getCurrentUser
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.User'
          description: Current user information
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: User not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Internal server error
      summary: Get current user
      tags:
      - authentication
  /auth/refresh:
    post:
      description: Extend the current session expiry time
      operationId: refreshSession
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SessionRefreshResponse'
          description: Session refreshed with new expiry
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Invalid or expired session
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.StandardErrorResponse'
          description: Failed to refresh session
      summary: Refresh user session
      tags:
      - authentication
  /campaigns:
    delete:
      description: Delete multiple campaigns in a single operation
      operationId: bulkDeleteCampaigns
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.BulkDeleteRequest'
        description: Bulk delete request with campaign IDs
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.BulkDeleteResult'
          description: Bulk delete results
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Bulk delete campaigns
      tags:
      - campaigns
    get:
      description: Retrieve a list of all campaigns with optional filtering and pagination
      operationId: listCampaigns
      parameters:
      - description: Maximum number of campaigns to return
        in: query
        name: limit
        schema:
          default: 20
          type: integer
      - description: Number of campaigns to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Filter by campaign status
        in: query
        name: status
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.APIResponse'
          description: List of campaigns
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Internal Server Error
      summary: List all campaigns
      tags:
      - campaigns
    post:
      description: Create a new campaign with specified configuration parameters
      operationId: createCampaign
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/services.CreateCampaignRequest'
        description: Campaign creation request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.CampaignOperationResponse'
          description: Campaign created successfully
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Create new campaign
      tags:
      - campaigns
...
  /personas:
    get:
      description: Retrieve a list of all personas with optional filtering by type
        and status
      parameters:
      - description: Maximum number of results
        in: query
        name: limit
        schema:
          default: 20
          type: integer
      - description: Number of results to skip
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      - description: Filter by enabled status
        in: query
        name: isEnabled
        schema:
          type: boolean
      - description: Filter by persona type (dns, http)
        in: query
        name: personaType
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/api.PersonaResponse'
                type: array
          description: List of personas
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid personaType parameter
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to list personas
      summary: List all personas
      tags:
      - personas
    post:
      description: Create a new persona (DNS or HTTP) with configuration details
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.CreatePersonaRequest'
        description: Persona creation request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PersonaResponse'
          description: Created persona
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Invalid request payload or configuration
        "409":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Persona with name and type already exists
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Failed to create persona
      summary: Create persona
      tags:
      - personas
...
  /personas/{id}/test:
    post:
      description: Test a persona configuration to verify it works correctly
      parameters:
      - description: Persona ID (UUID)
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PersonaTestResponse'
          description: Persona test results
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Persona not found
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Test persona
      tags:
      - personas
...
servers:
- url: /api/v2